chore: restore `source/numeric.d.ts`
