--- test-d/tsconfig.json
@@ -1,7 +1,0 @@
-{
-	"extends": "../tsconfig",
-	"exclude": [],
-	"compilerOptions": {
-		"strictNullChecks": true
-	}
-}
